change type of daily_cost field

+++ backend/hotel/models.py
@@ -4,7 +4,7 @@ from django.contrib.auth import get_user_model
 
 class Room(models.Model):
     room_name    = models.CharField(max_length=50)
-    daily_cost   = models.IntegerField()
+    daily_cost   = models.FloatField()
     num_of_seats = models.IntegerField()
 
     is_booked    = models.BooleanField(default=False)
